requirements.txt
streamlit
google-generativeai
python-dotenv
requests
pandas
plotly
pillow
numpy
PyPDF2
scikit-learn